make smaller database table cols

+++ src/database/migrations/2024_08_15_090228_create_image_variants_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             // Represents the maximum dimension (either width or height, whichever is larger) of the image variant.
             $table->string('size_type', 3); // e.g. max-allowed, desktop, mobile
             // Stores the quality of the image variant (e.g., for JPEG compression).
-            $table->integer('quality');
+            $table->tinyInteger('quality');
             // Stores the width of the image variant. Nullable because not all image variants may have an explicit width.
             $table->integer('width')->nullable();
             // Stores the height of the image variant. Nullable because not all image variants may have an explicit height.
